Move Pass rule test to separate file
