+++ program/tests/integration.rs
@@ -93,7 +93,7 @@ async fn test_payer_not_signer_fails() {
 }
 
 #[tokio::test]
-async fn test_additional_signer_and_amount() {
+async fn test_additional_signer_and_not_amount() {
     let mut context = program_test().start_with_context().await;
 
     // Create some rules.
@@ -206,51 +206,6 @@ async fn test_additional_signer_and_amount() {
 
     // Check that error is what we expect.
     assert_rule_set_error(err, RuleSetError::AmountCheckFailed);
-}
-
-#[tokio::test]
-async fn test_pass() {
-    let mut context = program_test().start_with_context().await;
-
-    // --------------------------------
-    // Create RuleSet
-    // --------------------------------
-    // Create a Pass Rule.
-    let pass_rule = Rule::Pass;
-
-    // Create a RuleSet.
-    let mut rule_set = RuleSet::new("test rule_set".to_string(), context.payer.pubkey());
-    rule_set
-        .add(Operation::Transfer.to_string(), pass_rule)
-        .unwrap();
-
-    println!("{:#?}", rule_set);
-
-    // Put the RuleSet on chain.
-    let rule_set_addr =
-        create_rule_set_on_chain(&mut context, rule_set, "test rule_set".to_string()).await;
-
-    // --------------------------------
-    // Validate Pass Rule
-    // --------------------------------
-    // Create a Keypair to simulate a token mint address.
-    let mint = Keypair::new().pubkey();
-
-    // Create a `validate` instruction.
-    let validate_ix = ValidateBuilder::new()
-        .rule_set_pda(rule_set_addr)
-        .mint(mint)
-        .additional_rule_accounts(vec![])
-        .build(ValidateArgs::V1 {
-            operation: Operation::Transfer.to_string(),
-            payload: Payload::default(),
-            update_rule_state: false,
-        })
-        .unwrap()
-        .instruction();
-
-    // Validate Transfer operation.
-    process_passing_validate_ix(&mut context, validate_ix, vec![]).await;
 }
 
 #[tokio::test]
